requirements.txt
streamlit>=1.32,<2.0
fpdf2>=2.7,<3.0
openai>=1.10,<2.0
python-dotenv>=1.0,<2.0
